Github Integration: Remove 'should be carried out'.
The paragraph "These items should already have been carried out:"
is wrong.

First, it is not necessarily true, it depends on the workflow
used. In particular, it is not true when applying the workflow
on an already otherwise working plugin.

Also, a note like this actually requires a reference. A reasonable
approach would be something like "these steps are done when
setting up the pristine plugin as described in the
xref::TBD.adoc[install page].

On too of this, the paragraph breaks the numbering of the steps.

~