Metadata push: Clarify steps, note on Windows script deadlock.
Divide the sometimes problematic step to run new-credentials
into two.

Remove the too strict language aboyt the python alias. It
is not strictly necessary, the stackoverflow question describes
how to make a permanent fix to avoid it. Likewise, that
Windows users "should" use git-bash is a little too strong.
For those who are really motivated it's possible to use
CMD.

Shorten the too lengthy explanation that "script must complete".
All scripts in a workflow must complete, this is implied. Focus
on how to handle when it not completes on Windows in a
prominent note.